Admin/requirements.txt
pypdf
langchain
streamlit
openai
faiss-cpu
boto3 #aws
langchain-community

